[DQL-211] Consolidate reporting functionality into reporting dir.
