+++ ckanext/data_qld/reporting_plugin.py
@@ -1,15 +1,20 @@
 import ckan.plugins as plugins
-import logic.action.get as get
+import ckanext.data_qld.reporting.logic.action.get as get
 import auth_functions as auth
 
-from ckanext.data_qld.logic import helpers
+from ckanext.data_qld.reporting.helpers import helpers
 
 
 class ReportingPlugin(plugins.SingletonPlugin):
+    plugins.implements(plugins.IConfigurer)
     plugins.implements(plugins.IActions, inherit=True)
     plugins.implements(plugins.IRoutes, inherit=True)
     plugins.implements(plugins.ITemplateHelpers, inherit=True)
     plugins.implements(plugins.IAuthFunctions)
+
+    # IConfigurer
+    def update_config(self, config_):
+        plugins.toolkit.add_resource('reporting/fanstatic', 'data_qld_reporting')
 
     # IActions
     def get_actions(self):
@@ -35,16 +40,16 @@ class ReportingPlugin(plugins.SingletonPlugin):
 
     def before_map(self, map):
         map.connect('/dashboard/reporting/export',
-                    controller='ckanext.data_qld.controllers.reporting:ReportingController',
+                    controller='ckanext.data_qld.reporting.controller:ReportingController',
                     action='export')
         map.connect('dashboard.reports', '/dashboard/reporting',
-                    controller='ckanext.data_qld.controllers.reporting:ReportingController',
+                    controller='ckanext.data_qld.reporting.controller:ReportingController',
                     action='index')
         map.connect('/dashboard/reporting/datasets/{org_id}/{metric}',
-                    controller='ckanext.data_qld.controllers.reporting:ReportingController',
+                    controller='ckanext.data_qld.reporting.controller:ReportingController',
                     action='datasets')
         map.connect('/dashboard/reporting/datarequests/{org_id}/{metric}',
-                    controller='ckanext.data_qld.controllers.reporting:ReportingController',
+                    controller='ckanext.data_qld.reporting.controller:ReportingController',
                     action='datarequests')
         return map
