+++ ckanext/data_qld/templates/reporting/index.html
@@ -3,6 +3,8 @@
 
 {% block primary_content_inner %}
 
+    {% resource "data_qld_reporting/reporting.css" %}
+
     <div>
         <h1>Engagement Report</h1>
